sahayak-backend/requirements.txt
# ================================
# SAHAYAK AI PLATFORM - ESSENTIAL DEPENDENCIES
# ================================

# ---- WEB FRAMEWORK ----
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
pydantic==2.5.0
python-dotenv==1.0.0

# ---- GOOGLE FREE AI SERVICE ----
google-generativeai

# ---- HTTP CLIENTS ----
httpx==0.25.2
requests==2.31.0

# ---- CORS MIDDLEWARE ----
fastapi-cors==0.0.6 